+++ apps/L1G/tinder/tinder-frontend/src/components/MultiStepForm2.tsx
@@ -3,8 +3,8 @@ import HowOldAreYou from '@/components/HowOldAreYou';
 import { ProfileImages } from '@/components/ProfileImages';
 import { YouAreAllSet } from '@/components/YouAreAllSet';
 import { AppRouterInstance } from 'next/dist/shared/lib/app-router-context.shared-runtime';
-import { UserData } from '@/app/(auth)/signup/page';
 import { YourDetailsPage } from './YourDetailsPage';
+import { UserData } from 'types/chat';
 
 type Props = {
   step: string;
