+++ apps/L1G/tinder/tinder-frontend/src/components/YourDetailsPage.tsx
@@ -1,6 +1,6 @@
-import { UserData } from '@/app/(auth)/signup/page';
 import TitleContainer from './TitleContainer';
 import { ProfileForm } from './FormNew';
+import { UserData } from 'types/chat';
 
 type YourDetailsPageProps = {
   onSuccess: () => void;
